app/requirements.txt
asgiref==3.8.1
brotli==1.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography
dj-database-url==3.0.1
Django==5.1
django-extensions==4.1
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
mysqlclient==2.2.4
numpy==1.26.4
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pyarrow==17.0.0
pydotplus==2.0.2
PyJWT==2.10.1
pylance==0.16.1
PyMySQL==1.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sqlparse==0.5.1
stripe==12.5.1
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
whitenoise==6.11.0
dj-database-url
python-decouple
psycopg2-binary

